+++ lib/layers.rb
@@ -11,6 +11,7 @@ require 'layers/errors'
 require 'layers/adapters'
 require 'layers/configuration'
 require 'layers/logger'
+require 'layers/instrumenter'
 
 require 'layers/dsl'
 require 'layers/base_layer'