+++ lib/layers/dsl.rb
@@ -3,6 +3,7 @@
 require 'layers/dsl/callback_defaults'
 require 'layers/dsl/class_callable'
 require 'layers/dsl/inputs'
+require 'layers/dsl/instrumented'
 require 'layers/dsl/null_listener'
 require 'layers/dsl/observers'